services/agent-service/requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0

# AI/LLM
anthropic==0.40.0
openai==1.58.1

# Database & Storage
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP & Async
httpx==0.27.2
aiofiles==24.1.0

# Code Analysis & Processing
gitpython==3.1.43
beautifulsoup4==4.12.3
lxml==5.3.0

# Data Processing
pandas==2.2.3
Pillow==10.4.0

# Utilities
PyYAML==6.0.2
jinja2==3.1.4
